+++ fst/layout/RaidDpLayout.cc
@@ -41,7 +41,7 @@ typedef long v2do __attribute__ ((vector_size (VECTOR_SIZE)));
 //------------------------------------------------------------------------------
 
 RaidDpLayout::RaidDpLayout (XrdFstOfsFile* file,
-                            int lid,
+                            unsigned long lid,
                             const XrdSecEntity* client,
                             XrdOucErrInfo* outError,
                             eos::common::LayoutId::eIoType io,
